refactor: organize templates into directory structure
