+++ src/NextGen/DonationForm/Blocks/DonationFormBlock/app/templates/elements/Html.tsx
@@ -1,0 +1,9 @@
+import {ElementProps} from '../index';
+
+interface HtmlProps extends ElementProps {
+    html: string;
+}
+
+export default function Html({html}: HtmlProps) {
+    return <div dangerouslySetInnerHTML={{__html: html}} />;
+}
