+++ src/NextGen/DonationForm/Blocks/DonationFormBlock/app/templates/fields/Email.tsx
@@ -1,0 +1,10 @@
+import {FieldProps} from '../index';
+
+export default function Email({label, inputProps}: FieldProps) {
+    return (
+        <label>
+            {label}
+            <input type="email" {...inputProps} />
+        </label>
+    );
+}
