+++ src/NextGen/DonationForm/Blocks/DonationFormBlock/app/templates/fields/Hidden.tsx
@@ -1,0 +1,5 @@
+import {FieldProps} from '../index';
+
+export default function Hidden({inputProps}: FieldProps) {
+    return <input type="hidden" {...inputProps} />;
+}
