+++ src/NextGen/DonationForm/Blocks/DonationFormBlock/app/templates/fields/Text.tsx
@@ -1,0 +1,10 @@
+import {FieldProps} from '../index';
+
+export default function Text({label, inputProps}: FieldProps) {
+    return (
+        <label>
+            {label}
+            <input type="text" {...inputProps} />
+        </label>
+    );
+}
