+++ src/NextGen/DonationForm/Blocks/DonationFormBlock/app/templates/fields/TextArea.tsx
@@ -1,0 +1,10 @@
+import {FieldProps} from '../index';
+
+export default function TextArea({label, inputProps}: FieldProps) {
+    return (
+        <label>
+            {label}
+            <textarea {...inputProps} />
+        </label>
+    );
+}
